Question 2
I updated question 2 in the VariablesDataTypesOperators class. I was
shocked you did not want us to handle null values in this question.
When I used the if statement and had an else of null the Test failed
which was actually cool to see that a test will fail if the class is
producing an unknown result that is not being tested.